discord-bot/requirements.txt
hikari              # discord framework
hikari[speedups]
uvloop; os_name != 'nt'
hikari-lightbulb    # command handler
hikari-miru         # modals and buttons

python-dotenv       # .env file support
aiosqlite           # database
aiohttp             # http client
aiohttp[speedups]   # speedups for aiohttp